abstracted out document routes
